chore: move service contract logic to service_contract.rs file
